fixed long int problem

+++ includes/ft_printf.h
@@ -66,7 +66,7 @@ void		is_it_double_specifier(char *str, char *tmp, int i);
 void		param_free(t_parameter *p);
 int			who_is_biggest_of_2(int a, int b);
 int			who_is_biggest_of_3(int a, int b, int c);
-void		ft_len_zero_handling(t_parameter *p, int n, int len);
+void		ft_len_zero_handling(t_parameter *p, long long n, int len);
 
 /*
 ** ft_print_str.c
